+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/UpstreamKitTest.groovy
@@ -9,9 +9,9 @@ import org.gradle.testkit.runner.TaskOutcome
  */
 class UpstreamKitTest extends AbstractKitTest {
 
-    public static final String GRADLE_VERSION = '8.11'
-    public static final String ANDROID_PLUGIN_VERSION = '8.8.0'
-    public static final String KOTLIN_PLUGIN_VERSION = '2.1.0'
+    public static final String GRADLE_VERSION = '8.14'
+    public static final String ANDROID_PLUGIN_VERSION = '8.9.2'
+    public static final String KOTLIN_PLUGIN_VERSION = '2.1.20'
 
     def "Check violation detection without cache task"() {
         setup: