+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/android/multiplatform/KotlinMultiplatformKitTest.groovy
@@ -65,6 +65,61 @@ class KotlinMultiplatformKitTest extends AbstractKitTest {
         ]
     }
 
+    def "Check kotlin multiplatform registration after animalsniffer"() {
+        setup:
+        build """
+            plugins {
+            id 'ru.vyarus.animalsniffer'
+                id 'org.jetbrains.kotlin.multiplatform' version '2.0.21'                
+            }
+            
+            kotlin {
+                jvm()
+                
+                sourceSets {
+                    commonMain.dependencies {
+                        implementation 'org.slf4j:slf4j-api:1.7.25'
+                    }                    
+                }
+            }
+            
+            animalsniffer {
+                ignoreFailures = true
+            }
+
+            repositories { mavenCentral()}
+            dependencies {
+                signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'                
+            }                 
+        """
+        fileFromClasspath('src/jvmMain/kotlin/invalid/Sample.kt', '/ru/vyarus/gradle/plugin/animalsniffer/kotlin/invalid/Sample.kt')
+//        debug()
+
+        when: "run task"
+        BuildResult result = run('check')
+
+        then: "task successful"
+        result.task(':check').outcome == TaskOutcome.SUCCESS
+
+        then: "found 2 violations"
+        result.output.contains("2 AnimalSniffer violations were found in 1 files")
+        result.output.replaceAll('\r', '').contains(
+                """[Undefined reference] invalid.(Sample.kt:11)
+  >> int Boolean.compare(boolean, boolean)
+
+[Undefined reference] invalid.(Sample.kt:16)
+  >> java.nio.file.Path java.nio.file.Paths.get(String, String[])
+""")
+
+        then: "report correct"
+        File file = file('/build/reports/animalsniffer/jvmMain.text')
+        file.exists()
+        file.readLines() == [
+                "invalid.Sample:11  Undefined reference: int Boolean.compare(boolean, boolean)",
+                "invalid.Sample:16  Undefined reference: java.nio.file.Path java.nio.file.Paths.get(String, String[])"
+        ]
+    }
+
     def "Check kotlin multiplatform jvm with java"() {
         setup:
         build """