+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Detect java support by `java` plugin presence instead of `java-base`. 
+  Required to avoid duplicate tasks registration when kotlin-multiplatform plugin registered after animalsniffer (#116)
+
 ### 2.0.0 (2025-01-25)
 * (breaking) Drop gradle 5 and 6 support
 * (breaking) Fail check when no signatures declared (required to catch early configuration errors)